[Feature #160941788] write tests for endpoints
