+++ app/resources/v2/resources/endpoints/user.py
@@ -19,18 +19,9 @@ class UserResource(Resource):
     parser.add_argument("admin", type=bool, required=True, help="Role cannot be left blank")
     parser.add_argument("password", type=str, required=True, help="Password canot be left blank")
 
-    def get(self):
-        """
-            This function fetchs all users from the database and returns
-        """
-        userObject = User(name=None,password=None)
-        json_data = userObject.getallUser()
 
-        return json_data
-
- 
     def post(self):
-        """
+        """from flask_httpauth import HTTPBasicAuth
             This function adds a user into the database and assigns them a role
         """
         data = UserResource.parser.parse_args()
@@ -75,3 +66,17 @@ class UserLogin(Resource):
         response = userObject.login()
         token = userObject.generate_auth_token()
         return jsonify ({"Token":token.decode('UTF-8')})
+
+class GetAllUsersResources(Resource):
+    """
+        This function is responsibe to Get all Users Registered
+    """
+    @auth.login_required
+    def get(self):
+        """
+            This function fetchs all users from the database and returns
+        """
+        userObject = User(name=None,password=None)
+        json_data = userObject.getallUser()
+
+        return json_data
